# restserver/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WhatsApp API DietSystem",
    "version": "1.0.0",
    "description": "API para gerenciar sessões WhatsApp, mensagens, filas, webhooks e diagnóstico. Não inclua chaves em payloads ou repositórios."
  },
  "servers": [{"url": "https://zap.dietsystem.com.br", "description": "Produção"}],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "description": "ADMIN_API_KEY"},
      "adminToken": {"type": "apiKey", "in": "header", "name": "admintoken"},
      "instanceToken": {"type": "apiKey", "in": "header", "name": "token"}
    },
...
    "schemas": {
      "Error": {"type": "object", "required": ["error"], "properties": {"error": {"type": "string"}}, "example": {"error": "instance is not connected/logged in"}},
      "Instance": {"type": "object", "properties": {"id": {"type": "string"}, "token": {"type": "string"}, "name": {"type": "string"}, "status": {"type": "string", "enum": ["disconnected", "connecting", "connected", "hibernated"]}, "owner": {"type": "string"}, "profileName": {"type": "string"}}},
      "TextMessage": {"type": "object", "required": ["number", "text"], "properties": {"number": {"type": "string", "example": "5565999999999"}, "text": {"type": "string", "example": "Olá! Sua consulta está confirmada."}, "async": {"type": "boolean", "default": false}, "idempotencyKey": {"type": "string"}}},
      "MediaMessage": {"type": "object", "required": ["number", "file"], "properties": {"number": {"type": "string"}, "type": {"type": "string", "enum": ["image", "video", "audio", "document"]}, "file": {"type": "string", "description": "URL, base64 ou data URI"}, "text": {"type": "string"}, "fileName": {"type": "string"}, "docName": {"type": "string"}, "async": {"type": "boolean"}, "idempotencyKey": {"type": "string"}}},
...
      "WebhookInput": {"type": "object", "required": ["url"], "properties": {"url": {"type": "string", "format": "uri"}, "secret": {"type": "string"}, "events": {"oneOf": [{"type": "string"}, {"type": "array", "items": {"type": "string"}}]}, "excludeMessages": {"type": "array", "items": {"type": "string"}}, "addUrlEvents": {"type": "boolean", "default": false}, "addUrlTypesMessages": {"type": "boolean", "default": false}, "enabled": {"type": "boolean"}}},
      "UazapiWebhookConfig": {"type": "object", "required": ["id", "enabled", "url", "events", "excludeMessages", "addUrlEvents", "addUrlTypesMessages"], "properties": {"id": {"type": "string", "format": "uuid"}, "enabled": {"type": "boolean"}, "url": {"type": "string", "format": "uri"}, "events": {"type": "array", "items": {"type": "string"}}, "excludeMessages": {"type": "array", "items": {"type": "string"}}, "addUrlEvents": {"type": "boolean"}, "addUrlTypesMessages": {"type": "boolean"}}}
    },
    "responses": {
      "BadRequest": {"description": "Requisição inválida", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "Unauthorized": {"description": "Chave ausente ou inválida", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
...
    }
  },
  "paths": {
...
    "/webhook": {
      "get": {"tags": ["Webhooks", "Compatibilidade UAZAPI"], "summary": "Ler webhook da instância por token ou validar handshake Cloud API", "security": [{"instanceToken": []}, {}], "parameters": [{"name": "token", "in": "header", "schema": {"type": "string"}}, {"name": "hub.mode", "in": "query", "schema": {"type": "string"}}, {"name": "hub.verify_token", "in": "query", "schema": {"type": "string"}}, {"name": "hub.challenge", "in": "query", "schema": {"type": "string"}}], "responses": {"200": {"description": "Array de configuração da instância ou challenge validado", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/UazapiWebhookConfig"}}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}, "403": {"description": "Verify token Cloud inválido"}}},
      "post": {"tags": ["Compatibilidade UAZAPI"], "summary": "Configurar webhook compatível", "security": [{"instanceToken": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WebhookInput"}, "example": {"url": "https://api.exemplo.com/webhooks/uazapi", "enabled": true, "events": ["messages", "connection"], "excludeMessages": ["wasSentByApi", "fromMeYes", "isGroupYes"]}}}}, "responses": {"200": {"description": "Array com o webhook persistido", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/UazapiWebhookConfig"}}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}}}
    },
    "/instances": {
      "get": {"tags": ["Instâncias"], "summary": "Listar instâncias", "security": [{"bearerAuth": []}], "responses": {"200": {"description": "Lista", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Instance"}}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}}},
      "post": {"tags": ["Instâncias"], "summary": "Criar instância", "security": [{"bearerAuth": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["name"], "properties": {"name": {"type": "string"}, "adminField01": {"type": "string"}, "webhookUrl": {"type": "string"}, "webhookSecret": {"type": "string"}}}, "example": {"name": "clinica-cuiaba", "adminField01": "unidade-01"}}}}, "responses": {"201": {"description": "Criada", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Instance"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}}}
    },
...
    "/instance/init": {"post": {"tags": ["Compatibilidade UAZAPI"], "summary": "Criar instância compatível", "security": [{"adminToken": []}], "requestBody": {"required": true, "content": {"application/json": {"example": {"name": "clinica-cuiaba", "adminField01": "unidade-01", "systemName": "DietSystem"}}}}, "responses": {"200": {"description": "Instância e token"}, "401": {"$ref": "#/components/responses/Unauthorized"}}}},
    "/instance/connect": {"post": {"tags": ["Compatibilidade UAZAPI"], "summary": "Gerar QR compatível", "security": [{"instanceToken": []}], "responses": {"200": {"description": "Instância e QR"}}}},
    "/instance/status": {"get": {"tags": ["Compatibilidade UAZAPI"], "summary": "Consultar status compatível", "security": [{"instanceToken": []}], "responses": {"200": {"description": "Estado da instância"}}}},
    "/instance/all": {"get": {"tags": ["Compatibilidade UAZAPI"], "summary": "Listar instâncias compatíveis", "security": [{"adminToken": []}], "responses": {"200": {"description": "Lista de instâncias"}}}},
    "/instance/disconnect": {"post": {"tags": ["Compatibilidade UAZAPI"], "summary": "Desconectar compatível", "security": [{"instanceToken": []}], "responses": {"200": {"description": "Hibernada"}}}},
    "/instance/hibernate": {"post": {"tags": ["Compatibilidade UAZAPI"], "summary": "Hibernar compatível", "security": [{"instanceToken": []}], "responses": {"200": {"description": "Hibernada"}}}},
    "/instance/resume": {"post": {"tags": ["Compatibilidade UAZAPI"], "summary": "Retomar compatível", "security": [{"instanceToken": []}], "responses": {"202": {"description": "Reconexão solicitada"}}}},
    "/instance/reset": {"post": {"tags": ["Compatibilidade UAZAPI"], "summary": "Reset compatível", "security": [{"instanceToken": []}], "responses": {"202": {"description": "Reset solicitado"}}}},
    "/instance": {"delete": {"tags": ["Compatibilidade UAZAPI"], "summary": "Excluir instância compatível", "security": [{"instanceToken": []}], "responses": {"200": {"description": "Excluída"}}}},
    "/send/text": {"post": {"tags": ["Compatibilidade UAZAPI"], "summary": "Enviar texto compatível", "security": [{"instanceToken": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TextMessage"}}}}, "responses": {"200": {"description": "Enviado"}, "202": {"description": "Enfileirado"}}}},
    "/send/media": {"post": {"tags": ["Compatibilidade UAZAPI"], "summary": "Enviar mídia compatível", "security": [{"instanceToken": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MediaMessage"}}}}, "responses": {"200": {"description": "Enviada"}, "202": {"description": "Enfileirada"}}}},
    "/message/async": {"get": {"tags": ["Compatibilidade UAZAPI"], "summary": "Consultar fila compatível", "security": [{"instanceToken": []}], "responses": {"200": {"description": "Estado da fila"}}}, "delete": {"tags": ["Compatibilidade UAZAPI"], "summary": "Cancelar fila compatível", "security": [{"instanceToken": []}], "responses": {"200": {"description": "Jobs cancelados"}}}}
  }
}